Renames loader package to config
