+++ main.go
@@ -2,46 +2,19 @@ package main
 
 import (
 	"fmt"
-	"github.com/goamz/goamz/aws"
-	"github.com/goamz/goamz/s3"
-	"github.com/victorcampos/harbor/loader"
-	"io/ioutil"
+	"github.com/victorcampos/harbor/config"
+	"github.com/victorcampos/harbor/downloader"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"strings"
 )
 
 func main() {
-	harborConfig, err := loader.LoadConfig()
+	harborConfig, err := config.Load()
 	checkError(err)
 
-	fileListLength := len(harborConfig.Files)
-
-	if fileListLength > 0 {
-		awsAuthentication, err := aws.EnvAuth()
-		checkError(err)
-
-		s3Connection := s3.New(awsAuthentication, aws.USEast)
-		bucket := s3Connection.Bucket(harborConfig.S3.Bucket)
-		checkError(err)
-
-		fmt.Printf("Downloading from bucket: %s\r\n", harborConfig.S3.Bucket)
-		fmt.Printf("Files to be downloaded: %d\r\n", fileListLength)
-		fmt.Printf("Downloading to: %s\r\n\r\n", harborConfig.DownloadPath)
-		for key, value := range harborConfig.Files {
-			outputFilePath := filepath.Join(harborConfig.DownloadPath, value.FileName)
-
-			fmt.Printf("Downloading file number %d of %d...\r\n", key+1, fileListLength)
-			fmt.Printf("File: %s\r\n", outputFilePath)
-
-			contents, err := bucket.Get(harborConfig.S3.BasePath + value.S3Path)
-			checkError(err)
-
-			err = ioutil.WriteFile(outputFilePath, contents, 0644)
-			checkError(err)
-		}
-	}
+	err = downloader.DownloadFromS3(harborConfig)
+	checkError(err)
 
 	commandListLength := len(harborConfig.Commands)
